+++ src/components/music-player/index.js
@@ -1,0 +1,3 @@
+import "./music-player.css";
+
+export { musicPlayerState, musicPlayerMethods } from "./music-player.js";